requirements.txt
scikit-learn
Pillow
numpy
joblib
pandas
